feat: 请求层内拓扑结构最后一个operation node的输出，作为该层的输出
fix bug

fix

+++ front-end/src/components/LineCharts/src/computed.js
@@ -53,7 +53,7 @@ export const linearScale = (
   
     let minValue = values.min;
     let maxValue = values.max;
-    maxValue = axis === 'y' ? parseInt(maxValue * 1.2): maxValue;
+    maxValue = axis === 'y' ? parseFloat(maxValue * 1.2): maxValue;
  
     const scale = scaleLinear()
       .rangeRound(axis === 'x' ? [0, size] : [size, 0])
